Add example scalar IIngredient, rename OreFluid, simplify item HashCode
